+++ termsize_nosysioctl.go
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: BSD-3-Clause
 // Project: https://github.com/woozymasta/flags
 
-//go:build plan9 || appengine || wasm || aix
-// +build plan9 appengine wasm aix
+//go:build plan9 || appengine || wasm
+// +build plan9 appengine wasm
 
 package flags